# Stub `Fcom::GitHelpers#repo` in all tests by default