This significantly speeds up tests by avoiding a slow, real call to `git remote ...`.

This change also makes our tests more reliable / consistent across different environments, where the git origin might not always be `davidrunger/fcom`).

# Before

https://travis-ci.org/github/davidrunger/fcom/builds/695213378

```
-----------------------
Spec Performance Report
-----------------------
0.0 secs - Fcom::VERSION is not nil
0.454 secs - Fcom::Parser#parse prints stuff
0.456 secs - Fcom::OptionsHelpers#days when a days option is provided returns that number of days
0.465 secs - Fcom::GitHelpers#repo returns a string in form "username/repo" representing the repo
0.497 secs - Fcom::Querier#query executes a #system call with the expected command
0.511 secs - Fcom::OptionsHelpers#days when a days option was not provided returns nil
```

# After

https://travis-ci.org/github/davidrunger/fcom/builds/695214531

```
-----------------------
Spec Performance Report
-----------------------
0.0 secs - Fcom::VERSION is not nil
0.001 secs - Fcom::OptionsHelpers#days when a days option was not provided returns nil
0.001 secs - Fcom::OptionsHelpers#days when a days option is provided returns that number of days
0.001 secs - Fcom::Querier#query executes a #system call with the expected command
0.001 secs - Fcom::Parser#parse prints stuff
0.005 secs - Fcom::GitHelpers#repo when Fcom::GitHelpers#repo is not stubbed when `git remote ...` indicates that the remote is davidrunger/fcom returns a string in form "username/repo" representing the repo
```

# Conclusion

**Much better!** 😄