@davidrunger 2020-06-05
Big shoutout to the [`test-prof` gem](https://github.com/palkan/test-prof) for making it very easy to diagnose the slowness that I was seeing in tests, specifically via its [`ruby-prof` integration](https://test-prof.evilmartians.io/#/./ruby_prof?id=profiling-with-rubyprof).